feat(chess): remove console application with new logic
